Move the firmware format options into the RPC module
