+++ probe-rs-tools/src/bin/probe-rs/cmd/attach.rs
@@ -2,9 +2,9 @@ use std::path::PathBuf;
 
 use time::UtcOffset;
 
-use crate::FormatOptions;
 use crate::cmd::run::{MonitoringOptions, NormalRunOptions};
 use crate::rpc::client::RpcClient;
+use crate::rpc::functions::format::FormatOptions;
 use crate::rpc::functions::monitor::MonitorMode;
 use crate::rpc::utils::run_loop::VectorCatchConfig;
 use crate::util::cli::{self, FileMetadata, parse_metadata, rtt_client};
