+++ probe-rs-tools/src/bin/probe-rs/cmd/download.rs
@@ -2,7 +2,7 @@ use std::path::PathBuf;
 
 use crate::rpc::client::RpcClient;
 
-use crate::FormatOptions;
+use crate::rpc::functions::format::FormatOptions;
 use crate::util::cli;
 use crate::util::common_options::BinaryDownloadOptions;
 use crate::util::common_options::ProbeOptions;
